// Gate-Pass-SPA/src/app/shared/models/dashboard/myGatePasses.ts
export interface MyGatePasses {
  usersId: number;
  requestsId: number;
  vehicleId: number;
  plateNumber: number;
  year: number;
  zPlateCodeNameEn: string;
  zPlateCodeNameAr: string;
  zVehicleModelNameEn: string;
  zVehicleModelNameAr: string;
  zVehicleMakerNameEn: string;
  zVehicleMakerNameAr: string;
  zOwnershipTypeColorClass: string;
  zOwnershipTypeTextColor: string;
  requestDate: Date;
  zRequestsStatusId: number;
  zRequestsStatusNameEn: string;
  zRequestsStatusNameAr: string;
  zRequestsStatusColorClass: string;
}